refactor(sonic_similarity): extract SonicSimilarityPlugin to provider.py
Final stage of the split: moves the SonicSimilarityPlugin class (~1300
lines) to provider.py, matching the spotify/tidal/yousee convention of a
thin __init__.py shim plus a dedicated provider module. __init__.py
collapses from 1542 to 274 lines (-82%) and now contains only setup,
get_config_entries, _collect_status_text, and a redundant-alias re-export
of SonicSimilarityPlugin (satisfies mypy's explicit-export requirement
without forcing test imports to change).
